+++ NEWS.md
@@ -8,6 +8,9 @@
 * Added support for individuals of known origin (i.e. those identified with great accuracy using 
 parentage-based tagging) in the mixtures (#12).
 
+* Allow user to specify the total weight on the symmetrical Dirichlet prior for the mixing
+proportions in infer_mixture().
+
 
 
 # rubias 0.1.0